basic test implemented on timetracker component that tests number of cards rendered with mount. Attempt to pass context through shallow mounted aborted
