+++ src/pages/index.test.tsx
@@ -3,6 +3,8 @@ import { shallow } from 'enzyme';
 import TimeTracker from '@home/TimeTracker';
 import Home from '.';
 
+const mockUseGoogleLogin = jest.fn();
+
 describe('Home page', () => {
   it('should render TimeTracker', () => {
     const wrapper = shallow(<Home />);
